requirements.txt
Django==4.0.3
python-dotenv==0.20.0
postgres==4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
django-guardian==2.4.0
drf-nested-routers==0.93.4
